Database: store statistics also for single-test runs
Formerly first test in a run was always omited from statistics, under
assumption that first test has unstable results and that we have to make
multiple runs every time. Consequently, this omission led to missing
statics object on runs with just single test, which might make sense for
non-dnsperf tests.

For runs with a single test in them, we now copy singleton test result
into run statistics, which makes the result available to the rest of
Perflab as usual.